build: add priorty

+++ src/scheduler/pt.rs
@@ -62,7 +62,7 @@ impl PTSchedulerInner {
             let priority = info.priority;
             self.queues[priority as usize].push_back(tid);
             info!("in push, info.rest_slice is {}", info.rest_slice);
-            info!("in push, queues[pri].len() is {}", self.queues[priority].len());
+            info!("in push, queues[pri].len() is {}", self.queues[priority as usize].len());
         }
         /* 
         {
